% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_signature_counts.R
\name{aggregate_signature_counts}
\alias{aggregate_signature_counts}
\title{Bin signatures and aggregate counts of signatures from an MFF file.}
\usage{
aggregate_signature_counts(dir, pattern = NULL, breaks = NULL,
  flanking_bases = 1, output_rda = NULL)
}
\arguments{
\item{dir}{The directory containing the MFF files.}

\item{breaks}{The breaks used for binning the distance from ends of the reads of the mutations.}

\item{flanking_bases}{The number of flanking bases. Defaults to 1.}

\item{output_rda}{If non-NULL, the output matrix is stored in the path provided (should end with .rda).}
}
\value{
Returns a matrix with rows being the samples (each MFF file), columns representing
the binned mutational signatures, where bins are determined by \code{breaks} in user input.
The cells contain counts of the number of mutational signatures observed in that MFF file.
}
\description{
For each file in directory \code{dir}, aggregates the data from MFF files
into bins determined by distance from ends of the read based on \code{breaks} provided.
This function is an Internal data processing function for \code{aRchaic_cluster} and related
fucntions.
}
\keyword{aggregate_counts}
