% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRchaic_class.R
\name{aRchaic_class}
\alias{aRchaic_class}
\title{Classification (model-based and non-model based) of aDNA samples using DNA damage patterns}
\usage{
aRchaic_class(folders, class_labs = NULL, class_method = c("SVM",
  "classtpx"), run_from = "class", run_index = NULL, breaks = c(-1, seq(1,
  20, 1)), flanking_bases = 1, normalize = TRUE, svm.control = list(),
  classtpx.control = list())
}
\arguments{
\item{folders}{a vector of folder names hosting the MFF files. Each folder may represent
MFF files from same source sequenced and analysed together. (see vignette for
examples)}

\item{class_labs}{The class labels for the samples. Takes values from 1 to K (where K is the
number of classes) and NA. The NA values correspond to test samples whereas
the others are training samples assigned to a class between 1 and K.}

\item{class_method}{The method used for performing supervised learning. The options are \code{SVM}
and \code{classtpx}.}

\item{run_from}{Can take one of two options - \code{class} and \code{start}.  For \code{start},
the function will perform all steps - processing the MFF files and then
classification. for \code{class}, the function will first check
if the processed files are present already from previous run or otherwise.
 If present, it will use the processed files to run the classification model.}

\item{run_index}{The index vector of files to be included for each folder in the vector \code{folders}.
Defaults to using all files in the folder.}

\item{breaks}{The breaks used for binning the distance from ends of the reads of the mutations, when
processing the MFF files.}

\item{flanking_bases}{The numbe rof flanking bases to the mutation considered. Defaults to 1.}

\item{normalize}{If SVM option is chosen by the user, he can run it on the original data (for \code{normalize = FALSE})
or on log CPM normalized data (for \code{normalize = TRUE})}

\item{svm.control}{Control parameters for the SVM function in e1071.}

\item{classtpx.control}{Control parameters for the supervised GoM model application.}
}
\value{
For the \code{start} option of \code{run_from}, the function processes the MFF files
in each folder, aggregates them into a matrix and saves it in the folder as a .rda file. Then
it fits SVM or classtpx to the aggregated data from one or more folders as determined by
the vector \code{folders} and generates the probabilities of belonging to each class for the test
samples (which correspond to those samples that have NA in \code{class_labs}).
}
\description{
Performs classification of aDNA samples using DNA damage patterns- mutation,
flanking base, distance from read end, strand and strand break information. We provide two classifi
cation options - SVM and classification GoM (classtpx). We fit soft classification
for both cases, thereby obtaining probabilities of belonging to the classes for each test sample.
}
\keyword{aRchaic_class}
