% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRchaic_class_beta.R
\name{aRchaic_class_beta}
\alias{aRchaic_class_beta}
\title{Classification (model-based and non-model based) of aDNA samples using
filtered DNA damage patterns}
\usage{
aRchaic_class_beta(folders, type = c("mutation", "mutation-flank",
  "mutation-pos", "mutation-flank-pos", "specific-mutation-pos", "wo-strand",
  "wo-strand-break"), class_labs = NULL, class_method = c("SVM",
  "classtpx"), run_from = "class", run_index = NULL, breaks = c(-1, seq(1,
  20, 1)), flanking_bases = 1, pattern = NULL, max_pos = 20,
  normalize = TRUE, svm.control = list(), classtpx.control = list())
}
\arguments{
\item{folders}{a vector of folder names hosting the MFF files. Each folder may represent
MFF files from same source sequenced and analysed together. (see vignette for
examples)}

\item{type}{The type of filtering mechanism used. The possible options are
- \code{mutation} : just the mutations (C->T, C->G, C->A, T->A, T->C and T->G)
- \code{mutation-flank}: the mutation and flanking base features
- \code{mutation-pos}: the mutation and position of the mutation from the nearest end of read.
- \code{mutation-flank-pos}: mutation + flanking base + position from nearest end of read
- \code{specific-mutation-pos}: specific mutation (say C->T) + position of that mutation from nearest end of read
- \code{wo-strand}: all features except strand feature
- \code{wo-strand-break}: all features except the strand break feature}

\item{class_labs}{The class labels for the samples. Takes values from 1 to K (where K is the
number of classes) and NA. The NA values correspond to test samples whereas
the others are training samples assigned to a class between 1 and K.}

\item{class_method}{The method used for performing supervised learning. The options are \code{SVM}
and \code{classtpx}.}

\item{run_from}{Can take one of two options - \code{class} and \code{start}.  For \code{start},
the function will perform all steps - processing the MFF files and then
classification. for \code{class}, the function will first check
if the processed files are present already from previous run or otherwise.
 If present, it will use the processed files to run the classification model.}

\item{run_index}{The index vector of files to be included for each folder in the vector \code{folders}.
Defaults to using all files in the folder.}

\item{breaks}{The breaks used for binning the distance from ends of the reads of the mutations, when
processing the MFF files. Defaults to assigning each base as one bin from
end of the read to 20 bases.}

\item{flanking_bases}{The numbe rof flanking bases to the mutation considered. Defaults to 1.}

\item{pattern}{type of mutation (say "C->T"). Comes into play when \code{type} is \code{specific-mutation-pos}.
Defaults to NULL.}

\item{max_pos}{The maximum distance from ends of read that a mismatch would be considered.
The rest will be filtered out.}

\item{normalize}{If SVM option is chosen by the user, he can run it on the original data (for \code{normalize = FALSE})
or on log CPM normalized data (for \code{normalize = TRUE})}

\item{svm.control}{Control parameters for the SVM function in e1071.}

\item{classtpx.control}{Control parameters for the supervised GoM model application.}
}
\value{
For the \code{start} option of \code{run_from}, the function processes the MFF files
in each folder, aggregates them into a matrix and saves it in the folder as a .rda file.
Then it filters the signatures based on the \code{type} selected. Subsequently,
it fits SVM or classtpx to the aggregated data from one or more folders as determined by
the vector \code{folders} and generates the probabilities of belonging to each class for the test
samples (which correspond to those samples that have NA in \code{class_labs}).
}
\description{
Performs classification of aDNA samples using filtered DNA damage patterns,
which comprises of some subset of the 5 features (mutation,
flanking base, distance from read end, strand and strand break) but not all of them.
We provide two classification options - SVM and classification GoM (classtpx). We fit
soft classification for both cases, thereby obtaining probabilities of belonging to the
classes for each test sample.
}
\keyword{aRchaic_class_beta}
