% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRchaic_cluster.R
\name{aRchaic_cluster}
\alias{aRchaic_cluster}
\title{Grade of Membership (GoM) model clustering of aDNA samples using DNA damage patterns}
\usage{
aRchaic_cluster(folders, K, tol = 100, labs = NULL, levels = NULL,
  run_from = c("start", "gom", "plot"), run_index = NULL, breaks = c(-1,
  seq(1, 20, 1)), flanking_bases = 1, gom_method = "independent",
  topic_cols = c("red", "blue", "darkgoldenrod1", "cyan", "firebrick",
  "green", "hotpink", "burlywood", "yellow", "darkgray", "deepskyblue",
  "darkkhaki", "brown4", "darkorchid", "magenta", "yellow", "azure1", "azure4"),
  structure.control = list(), logo.control = list(),
  topics.control = list(), output_dir = NULL, structure_width = 5,
  structure_height = 8)
}
\arguments{
\item{folders}{a vector of folder names hosting the MFF files. Each folder may represent
MFF files from same source sequenced and analysed together. (see vignette for
examples)}

\item{K}{the number of clusters to fit to the model.}

\item{tol}{The tolerance level of convergence of the GoM model fit}

\item{labs}{The labels used to group the samples in visualization. May be used to distinguish
samples from different labs, or different library prep.}

\item{levels}{The levels of the \code{labs} vector. Defaults to all unique values of \code{labs} vector.}

\item{run_from}{Can take one of three terms - \code{start}, \code{gom} and \code{plot}. For \code{start},
the function will perform all steps - processing the MFF files, clustering
and visualization from scratch. For \code{gom}, the function will first check
if the processed files are present already from previous run or from aRchaic_pool().
If present, it will use the processed files to run clustering and then do subsequent
visualization. For \code{plot}, it will first check if there is a clustering output
already present from previous clustering fit, if present, it will use it to do the
visualization. If not, will switch to \code{gom}. \code{plot} and \code{gom} are designed
to save time if user is repeating the run with the same data with updates to the plot or to
model respectively.}

\item{run_index}{The index vector of files to be included for each folder in the vector \code{folders}.
Defaults to using all files in the folder.}

\item{breaks}{The breaks used for binning the distance from ends of the reads of the mutations, when
processing the MFF files. Defaults to assigning each base as one bin from
end of the read to 20 bases.}

\item{flanking_bases}{The numbe rof flanking bases to the mutation considered. Defaults to 1.}

\item{gom_method}{The GoM method type. Defaults to \code{independent} model proposed by
Y. Shiraichi and M. Stephens. The other option is to use the \code{full}
model which is uses the \code{maptpx} package by Matt Taddy.}

\item{topic_cols}{colors attributed to each cluster used for the Structure plot visualization
of the grades of membership.}

\item{structure.control}{Control parameters for the Structure plot representation in
\code{StructureGGplot()} fucntion.}

\item{logo.control}{Control parameters for the logo plot representation \code{damageLogo5()}
function.}

\item{topics.control}{Control parameters for the maptpx GoM model fit.}

\item{output_dir}{The output directory where the model, Structure plot and the logo plots
of the clusters will be saved. If NULL, it picks the current working directory.}

\item{structure_width}{The width of the image of the Structure plot representation.}

\item{structure_height}{The height of the image of the Structure plot representation.}
}
\value{
For the \code{start} option of \code{run_from}, the function processes the MFF files
in each folder, aggregates them into a matrix and saves it in the folder as a .rda file. Then
upon GoM model fitting, it saves model output in a model.rda file, the Structure plot in a
structure.png file and the logo plots in \code{logo_clus_k.png} files, with \code{k}
representing the cluster k -  all in the \code{output_dir} provided. For \code{gom} and
\code{plot} options of the function, each or both rda generation steps may be skipped
respectively, if they are already present.
}
\description{
Performs GoM model clustering of aDNA samples using DNA damage patterns- mutation,
flanking base, distance from read end, strand and strand break information. Upon performing the
model fit, it performs model visualizations and saves the figures in a user defined directory.
}
\keyword{aRchaic_cluster}
