% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRchaic_cluster_beta.R
\name{aRchaic_cluster_beta}
\alias{aRchaic_cluster_beta}
\title{Grade of Membership (GoM) model clustering of aDNA samples using filtered DNA damage patterns}
\usage{
aRchaic_cluster_beta(mat, K, type = c("mutation", "mutation-flank",
  "mutation-pos", "mutation-flank-pos", "specific-mutation-pos", "wo-strand",
  "wo-strand-break"), tol = 0.01, pattern = NULL, max_pos = 20,
  labs = NULL, levels = NULL, flanking_bases = 1,
  gom_method = "independent", topic_cols = c("red", "blue",
  "darkgoldenrod1", "cyan", "firebrick", "green", "hotpink", "burlywood",
  "yellow", "darkgray", "deepskyblue", "darkkhaki", "brown4", "darkorchid",
  "magenta", "yellow", "azure1", "azure4"), structure.control = list(),
  graph.control = list(), topics.control = list(), output_dir = NULL,
  structure_width = 5, structure_height = 8, inflation = rep(2, 1, 2),
  output_width = 1200, output_height = 700)
}
\arguments{
\item{mat}{A matrix of signature counts of samples obtained from aRchaic_pool() on one or more.
The signatures should contain al features}

\item{K}{the number of clusters to fit to the model.}

\item{type}{The type of filtering mechanism used. The possible options are
- \code{mutation} : just the mutations (C->T, C->G, C->A, T->A, T->C and T->G)
- \code{mutation-flank}: the mutation and flanking base features
- \code{mutation-pos}: the mutation and position of the mutation from the nearest end of read.
- \code{mutation-flank-pos}: mutation + flanking base + position from nearest end of read
- \code{specific-mutation-pos}: specific mutation (say C->T) + position of that mutation from nearest end of read
- \code{wo-strand}: all features except strand feature
- \code{wo-strand-break}: all features except the strand break feature}

\item{tol}{The tolerance level of convergence of the GoM model fit}

\item{pattern}{type of mutation (say "C->T"). Comes into play when \code{type} is \code{specific-mutation-pos}.
Defaults to NULL.}

\item{max_pos}{The maximum distance from ends of read that a mismatch would be considered.
The rest will be filtered out.}

\item{labs}{The labels used to group the samples in visualization. May be used to distinguish
samples from different labs, or different library prep.}

\item{levels}{The levels of the \code{labs} vector. Defaults to all unique values of \code{labs} vector.}

\item{flanking_bases}{The numbe rof flanking bases to the mutation considered. Defaults to 1.}

\item{gom_method}{The GoM method type. Defaults to \code{independent} model proposed by
Y. Shiraichi and M. Stephens. The other option is to use the \code{full}
model which is uses the \code{maptpx} package by Matt Taddy.}

\item{topic_cols}{colors attributed to each cluster used for the Structure plot visualization
of the grades of membership.}

\item{structure.control}{Control parameters for the Structure plot representation in
\code{StructureGGplot()} fucntion. Defaults to \code{list()}}

\item{graph.control}{Control parameters for the graphical representation of the clusters.
Defaults to \code{list()}.}

\item{topics.control}{Control parameters for the maptpx GoM model fit.}

\item{output_dir}{The output directory where the model, Structure plot and the logo plots
of the clusters will be saved. If NULL, it picks the current working directory.}

\item{structure_width}{The width of the image of the Structure plot representation.}

\item{structure_height}{The height of the image of the Structure plot representation.}

\item{inflation}{The inflation of flanking base logo plot if needed. Defaults to c(2,1,2),
which means the flanking base heights will be inflated 2 times with respect
to mutation in the middle in the logo plot.}

\item{output_width}{the width of the output plots of the graphs for cluster representation.}

\item{output_height}{the height of the output plots of the graphs for cluster representation.}
}
\value{
The function first filters the data based on the filtering \code{type} provided. Then
it performs GoM model fitting on the filtered data and saves model output in a model.rda file
Then it performs visualization of the grades of memebrships of the clusters and graphs for
cluster representation and save them as  Structure plot in a
structure.png file and the logo plots in \code{logo_clus_k.png} files, with \code{k}
representing the cluster k respectively-  all in the \code{output_dir} provided.
}
\description{
Performs GoM model clustering of aDNA samples using filtered DNA damage
patterns, which comprises of some subset of the 5 features (mutation,
flanking base, distance from read end, strand and strand break). An example is mutation-flank,
which would mean we just used mutation and flanking base features to do the clustering.
Upon performing the model fit, it performs model visualizations and saves the figures
in a user defined directory.
}
\keyword{aRchaic_cluster_beta}
