% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRchaic_mutation_trend.R
\name{aRchaic_mutation_trend}
\alias{aRchaic_mutation_trend}
\title{Mutation trends along the read at the  5' and 3' ends.}
\usage{
aRchaic_mutation_trend(file, pattern = c("C->T", "C->A", "C->G", "T->A",
  "T->C", "T->G", "G->A", "G->C", "G->T", "A->C", "A->G", "A->C"),
  plot_type = c("left", "right"), sample_name = NULL, strand = "both",
  numBases = 20, cols = c("black", "yellow", "green", "blue", "orange",
  "purple", "red", "pink", "brown", "gray", "darkseagreen", "magenta"),
  legend_cex = 0.5, main_cex = 0.7, lab_cex = 0.7,
  layout_left = "topright", layout_right = "topleft")
}
\arguments{
\item{file}{a MFF file.}

\item{pattern}{a vector of mutation types for which the trends are to be plotted.}

\item{plot_type}{if "left", the trend at the 5' end is calculated, else at the 3' end.}

\item{sample_name}{The name of the sample corresponding to the file. Defaults to filename}

\item{strand}{can take either +, - or both, representing whether the function consider positive,
negative or both strand mutations to calculate the trends.}

\item{numBases}{The number of bases from the end upto which the trend is to be plotted.
Defaults to 20.}

\item{cols}{Colors of the mutation types in the plot.}

\item{legend_cex}{The size of the legend. Defaults to 0.5}

\item{main_cex}{The size of the title of the figure. Defaults to 0.7.}

\item{lab_cex}{The size of the labels of the figure. Defaults to 0.5.}

\item{layout_left}{The layout of the legend if the plot is at 5' end. Defaults to "topright".}

\item{layout_right}{The layout of the legend if the plot is at 3' end. Defaults to "topleft".}
}
\value{
The function creates a logo plot representation of the mutational features for a single
        MFF file.
}
\description{
Plots mutation trends along the read at the 5' or at the 3' end
}
\keyword{aRchaic_mutation_trend}
