% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRchaic_pca.R
\name{aRchaic_pca}
\alias{aRchaic_pca}
\title{Performs Principal Component Analysis (PCA) of MFF samples using aDNA damage
patterns}
\usage{
aRchaic_pca(folders, labs = NULL, run_from = c("start", "pca", "plot"),
  run_index = NULL, breaks = c(-1, seq(1, 20, 1)), flanking_bases = 1,
  normalize = TRUE, cols = c("red", "blue", "darkgoldenrod1", "cyan",
  "firebrick", "green", "hotpink", "burlywood", "darkkhaki", "yellow",
  "darkgray", "deepskyblue", "brown4", "darkorchid", "magenta", "azure1",
  "azure4"), pcs_to_plot = c("PC1", "PC2", "PC3"), lay = NULL,
  plot_width = 10, plot_height = 7, output_dir = NULL)
}
\arguments{
\item{folders}{a vector of folder names hosting the MFF files. Each folder may represent
MFF files from same source sequenced and analysed together. (see vignette for
examples)}

\item{labs}{The labels used to group the samples in visualization. May be used to distinguish
samples from different labs, or different library prep.}

\item{run_from}{Can take one of three terms - \code{start}, \code{pca} and \code{plot}. For \code{start},
the function will perform all steps - processing the MFF files, PCA fitting
and visualization from scratch. For \code{pca}, the function will first check
if the processed files are present already from previous run or from aRchaic_pool().
If present, it will use the processed files to run PCA, save the model
as \code{pca.rda} file in the \code{output_dir} and then do subsequent
visualization. For \code{plot}, it will first check if there is a
\code{pca.rda} in \code{output_dir}. If present, it will use it to do the
visualization. If not, will switch to \code{pca}. \code{plot} and \code{pca} are
designed to save time if user is repeating the run with the same data with
updates to the plot or to model respectively.}

\item{run_index}{The index vector of files to be included for each folder in the vector \code{folders}.
Defaults to using all files in the folder.}

\item{breaks}{The breaks used for binning the distance from ends of the reads of the mutations, when
processing the MFF files.}

\item{flanking_bases}{The numbe rof flanking bases to the mutation considered. Defaults to 1.}

\item{normalize}{If SVM option is chosen by the user, he can run it on the original data (for \code{normalize = FALSE})
or on log CPM normalized data (for \code{normalize = TRUE})}

\item{cols}{colors attributed to each label in the \code{labs} vector.}

\item{pcs_to_plot}{a vector of the indicators of which PCs to plot. For example
\code{pcs_to_plot = c("PC1", "PC2", "PC3")} will plot three scatter
plots of PC1 vs PC2, PC1 vs PC3 and PC2 vs PC3.}

\item{lay}{A layout matrix. Defaults to to the upper diagonal layout. For other
options, check \code{grid.arrange} page
(https://cran.r-project.org/web/packages/gridExtra/vignettes/arrangeGrob.html).}

\item{plot_width}{The width of the PCA plot figure. Defaults to 10.}

\item{plot_height}{The height of the PCA plot figure. Defaults to 7.}

\item{output_dir}{The output directory where the model and the PCA plot figure are saved.
If NULL, it picks the current working directory.}
}
\value{
For the \code{start} option of \code{run_from}, the function processes the MFF files
in each folder, aggregates them into a matrix and saves it in the folder as a .rda file. Then
it fits PCA to the aggregated data from one or more folders as determined by
the vector \code{folders} and saves the result as \code{pca.rda} file in the
\code{output_dir}. Then the PCA plot corresponding to PCS given in \code{pcs_to_plot} is saved
as \code{pca.png} file in the \code{output_dir}.For option \code{pca}, the function skips
the data aggregation and processing step if the .rda files in the folders are present. For
\code{plot}, the functions skips the PCA fitting if the \code{pca.rda} file is already
present in \code{output_dir}. The \code{pca} and \code{plot} options are meant to reduce
time on repeated jobs.
}
\description{
Performs principal component analysis on the aDNA samples using the mutational
signature patterns- mutation, flanking base, distance from read end, strand and
strand break information.
}
\keyword{aRchaic_pca}
