% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archaic_pool.R
\name{archaic_pool}
\alias{archaic_pool}
\title{Aggregating signature counts from MFF files from one or more studies
(folders).}
\usage{
archaic_pool(dat)
}
\arguments{
\item{dat}{may be a list of count data matrices generated by archaic_prepare()
or a vector of folder names containing MFF files.}
}
\value{
The function creates a matrix of counts with combined samples from
multiple studies (stored as .csv MFF files and/or .RData file). The number of
rows of this matrix is same as the number of study samples across all the
studies considered, with columns representing mismatch signatures and the c
matrix cells recording the counts of the signatures occurring in the samples.
}
\description{
Aggregate signature counts data using archaic_prepare() from
multiple study directories into a single data frame or matrix.
}
\keyword{archaic_pool}
