% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRchaic_tsne.R
\name{aRchaic_tsne}
\alias{aRchaic_tsne}
\title{Performs non-linear dimension reduction using t-SNE of MFF samples using aDNA damage
patterns}
\usage{
aRchaic_tsne(folders, labs = NULL, run_from = c("start", "tsne", "plot"),
  run_index = NULL, breaks = c(-1, seq(1, 20, 1)), flanking_bases = 1,
  normalize = TRUE, cols = c("red", "blue", "darkgoldenrod1", "cyan",
  "firebrick", "green", "hotpink", "burlywood", "darkkhaki", "yellow",
  "darkgray", "deepskyblue", "brown4", "darkorchid", "magenta", "azure1",
  "azure4"), dims = 3, lay = NULL, filter_indices = NULL,
  plot_width = 10, plot_height = 7, output_dir = NULL)
}
\arguments{
\item{folders}{a vector of folder names hosting the MFF files. Each folder may represent
MFF files from same source sequenced and analysed together. (see vignette for
examples)}

\item{labs}{The labels used to group the samples in visualization. May be used to distinguish
samples from different labs, or different library prep.}

\item{run_from}{Can take one of three terms - \code{start}, \code{tsne} and \code{plot}. For \code{start},
the function will perform all steps - processing the MFF files, t-SNE fitting
and visualization from scratch. For \code{tsne}, the function will first check
if the processed files are present already from previous run or from aRchaic_pool().
If present, it will use the processed files to run t-SNE, save the model
as \code{tsne.rda} file in the \code{output_dir} and then do subsequent
visualization. For \code{plot}, it will first check if there is a
\code{tsne.rda} in \code{output_dir}. If present, it will use it to do the
visualization. If not, will switch to \code{tsne}. \code{plot} and \code{tsne} are
designed to save time if user is repeating the run with the same data with
updates to the plot or to model respectively.}

\item{run_index}{The index vector of files to be included for each folder in the vector \code{folders}.
Defaults to using all files in the folder.}

\item{breaks}{The breaks used for binning the distance from ends of the reads of the mutations, when
processing the MFF files.}

\item{flanking_bases}{The numbe rof flanking bases to the mutation considered. Defaults to 1.}

\item{normalize}{If SVM option is chosen by the user, he can run it on the original data (for \code{normalize = FALSE})
or on log CPM normalized data (for \code{normalize = TRUE})}

\item{cols}{colors attributed to each label in the \code{labs} vector.}

\item{dims}{The dimensions of the manifold on which the data is projected.}

\item{lay}{A layout matrix. Defaults to to the upper diagonal layout. For other
options, check \code{grid.arrange} page
(https://cran.r-project.org/web/packages/gridExtra/vignettes/arrangeGrob.html).}

\item{filter_indices}{An option to select a subset of features to run tSNE. t-SNE is expensive
on around 15,000 features, so this can be a easy way to test the model with
fewer features. Defaults to NULL, meaning all features are taken.}

\item{plot_width}{The width of the t-SNE plot figure. Defaults to 10.}

\item{plot_height}{The height of the t-SNE plot figure. Defaults to 7.}

\item{output_dir}{The output directory where the model and the t-SNE plot figure are saved.
If NULL, it picks the current working directory.}
}
\value{
For the \code{start} option of \code{run_from}, the function processes the MFF files
in each folder, aggregates them into a matrix and saves it in the folder as a .rda file. Then
it fits t-SNE to the aggregated data from one or more folders as determined by
the vector \code{folders} and saves the result as \code{tsne.rda} file in the
\code{output_dir}. Then the t-SNE plot corresponding to the number of dimensions
to fit \code{dims} is saved as \code{tsne.png} file in the \code{output_dir}.For option \code{tsne}, the function skips
the data aggregation and processing step if the .rda files in the folders are present. For
\code{plot}, the functions skips the t-SNE fitting if the \code{tsne.rda} file is already
present in \code{output_dir}. The \code{tsne} and \code{plot} options are meant to reduce
time on repeated jobs.
}
\description{
Performs t-SNE on the aDNA samples using the mutational
signature patterns- mutation, flanking base, distance from read end, strand and
strand break information.
}
\keyword{aRchaic_tsne}
