% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRchaic_view.R
\name{aRchaic_view}
\alias{aRchaic_view}
\title{Logo plot representation of an MFF file.}
\usage{
aRchaic_view(file, breaks = c(-1, seq(1, 20, 1)), flanking_bases = 1,
  logo.control = list(), title = NULL, output_dir = NULL,
  filename = "logo")
}
\arguments{
\item{file}{a MFF file.}

\item{breaks}{The breaks used for binning the distance from ends of the reads of the mutations, when
processing the MFF files.}

\item{flanking_bases}{The numbe rof flanking bases to the mutation considered. Defaults to 1.}

\item{logo.control}{The control parameters for the logo plot representation of the sample.
Check the input parameters of \code{damageLogo5()} for the control parameters.}

\item{title}{The name of the sample. If not provided, defaults to the filename.}

\item{output_dir}{The directory where the logo plot is saved.}

\item{filename}{The name of the saved logo image file. Defaults to "logo".}
}
\value{
The function creates a logo plot representation of the mutational features for a single
        MFF file.
}
\description{
Aggregate the signatures into bins for an MFF file and then based on the counts
of mutational signatures, creates a logo plot representation of the mutational patterns in the
sample represented by the MFF file.
}
\keyword{aRchaic_view}
