% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{base_probs_moderns}
\alias{base_probs_moderns}
\title{A modern background probability distribution on the features - mismatch,
flanking bases, strand breaks, position
A list comprising of probability distribution on features such as mismatches,
flanking bases, strand break bases and position of mismatch from the end
of the read.}
\format{
A list
}
\usage{
base_probs_moderns
}
\description{
A modern background probability distribution on the features - mismatch,
flanking bases, strand breaks, position
A list comprising of probability distribution on features such as mismatches,
flanking bases, strand break bases and position of mismatch from the end
of the read.
}
\keyword{datasets}
