% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/club_signature_counts.R
\name{club_signature_counts}
\alias{club_signature_counts}
\title{Club signatures in an MFF file to account for strand effect}
\usage{
club_signature_counts(signature_counts, flanking_bases = 1)
}
\arguments{
\item{signature_counts}{The matrix of counts of all signatures as produced by
\code{aggregate_bin_counts()}.}

\item{flanking_bases}{The number of flanking bases to the mutation. Defaults to 1.}
}
\value{
Returns a matrix of clubbed signatures. The default choice of the mutation
signatures are C->T, C->A, C->G, T->A, T->C and T->G. Other mutations are clubbed to
above signatures (for example G->A is clubbed with C->T).
}
\description{
Function for clubbing signatures from the two strands to remove
strand bias. An example is G->A and C->T are clubbed into C->T.
}
\keyword{club_signature_counts}
