% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damageLogo_5.R
\name{damageLogo_five}
\alias{damageLogo_five}
\title{Builds damage logo plots based on full mutation signatures (mutation, flanking base, position, strand and strand break)}
\usage{
damageLogo_five(theta_pool, sig_names = NULL, ic.scale = TRUE,
  max_pos = 20, flanking_bases = 1, yscale_change = TRUE, xaxis = TRUE,
  yaxis = TRUE, xlab = " ", xaxis_fontsize = 10, xlab_fontsize = 20,
  title_aligner = 15, y_fontsize = 20, title_fontsize = 20,
  mut_width = 2, start = 1e-04, renyi_alpha = 1, inflation_factor = c(2,
  1, 2), pop_names = paste0("Cluster ", 1:dim(theta_pool)[2]),
  logoport_x = 0.25, logoport_y = 0.5, logoport_width = 0.28,
  logoport_height = 0.4, lineport_x = 0.9, lineport_y = 0.4,
  lineport_width = 0.32, lineport_height = 0.28, breaklogoport_x = 1,
  breaklogoport_y = 0.4, breaklogoport_width = 0.4,
  breaklogoport_height = 0.5, barport_x = 0.58, barport_y = 0.6,
  barport_width = 0.25, barport_height = 0.4, output_dir = NULL,
  output_width = 1200, output_height = 700)
}
\arguments{
\item{theta_pool}{The theta matrix obtained from running the grade of membership model that stores for each cluster, the
probability distribution over all the mutational signatures.}

\item{sig_names}{The mutational signature names. Defaults to the rownames of the theta matrix above.}

\item{ic.scale}{A binary variable indicating whether the height of the bars for substitution and flanking bases should be
adjusted by the information criterion.}

\item{max_pos}{The maximum distance from the end of the read upto which mutations are considered.}

\item{flanking_bases}{The number of flanking bases of the mutational signature.}

\item{yscale_change}{A binary variable indicating whether the Y axis scale should be adjusted based on the size of the
logos, defaults to TRUE.}

\item{xaxis}{A binary indicating whether the X axis of the logo plot should be shown}

\item{yaxis}{A binary indicating whether the Y axis of the logo plot should be shown}

\item{xlab}{The labels for X axis.}

\item{xaxis_fontsize}{The fontsize of the X axis ticks.}

\item{xlab_fontsize}{The fontsize of the X axis labels.}

\item{y_fontsize}{The fontsize of the Y axis ticks.}

\item{mut_width}{Thw width of the bar for the mutation at the center.}

\item{start}{The starting point of the stacking of logos on the Y axis. Should be close to 0, defau;ts to 0.0001.}

\item{renyi_alpha}{The entropy scale for the Renyi entropy on the flanking bases and mutations.}

\item{pop_names}{The title of the plot. Defaults to the cluster labels.}

\item{logoport_x}{the X-axis position of the plot window for the logo plot}

\item{logoport_y}{the Y-axis position of the plot window for the logo plot}

\item{logoport_width}{the width of the plot window for the logo plot}

\item{logoport_height}{the width of the plot window for the logo plot}

\item{lineport_x}{the X-axis position of the plot window for the mutational profile line plot.}

\item{lineport_y}{the Y-axis position of the plot window for the mutational profile line plot.}

\item{lineport_width}{the width of the plot window for the mutational profile line plot.}

\item{lineport_height}{the width of the plot window for the mutational profile line plot.}

\item{breaklogoport_x}{the X-axis position of the plot window for strand break logo plot.}

\item{breaklogoport_y}{the Y-axis position of the plot window for the strand break logo plot.}

\item{breaklogoport_width}{the width of the plot window for the strand break logo plot.}

\item{breaklogoport_height}{the width of the plot window for the strand break logo plot.}

\item{barport_x}{the X-axis position of the plot window for strand bar plot.}

\item{barport_y}{the Y-axis position of the plot window for the strand bar plot.}

\item{barport_width}{the width of the plot window for the strand bar plot.}

\item{barport_height}{the width of the plot window for the strand bar plot.}

\item{output_dir}{The directory where the logo plot will be saved.}

\item{output_width}{The width of the logo plot figure.}

\item{output_height}{the height of the logo plot figure.}

\item{inflation}{The inflation scale of flanking bases with respect to mutation.
Will be a 3 length vector. Defaults to c(1,1,1) implying no inflation.
c(2,1,2) will mean the flanking bases are 2 times inflated compared to mutation.}
}
\value{
Returns logo plots for each cluster

Returns logo plot for each cluster
}
\description{
Damage Logo plots for each cluster from the GoM model fit. It showcases the
different mutational features - for example, mutation, flanking base, position on read,
strand and strand break information for each cluster.
}
