% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damageLogo_6.R
\name{damageLogo_six}
\alias{damageLogo_six}
\title{Builds damage logo plots based on full mutation signatures (mutation, flanking base, position, strand and strand break)}
\usage{
damageLogo_six(theta_pool, sig_names = NULL, max_pos = 20,
  flanking_bases = 1, mutlogo.control = list(),
  breaklogo.control = list(), base_probs_list = NULL, clip = 0,
  mut_ranges = c(0, 0), break_ranges = c(0, 0), logoport_x = 0.7,
  logoport_y = 0.5, logoport_width = 1.2, logoport_height = 1.1,
  breaklogoport_x = 0.5, breaklogoport_y = 0.4, breaklogoport_width = 0.7,
  breaklogoport_height = 1, lineport_x = 0.4, lineport_y = 0.5,
  lineport_width = 1, lineport_height = 1, output_dir = NULL,
  output_width = 1200, output_height = 700)
}
\arguments{
\item{theta_pool}{The theta matrix obtained from running the grade of membership model that stores for each cluster, the
probability distribution over all the mutational signatures.}

\item{sig_names}{The mutational signature names. Defaults to the rownames of the theta matrix above.}

\item{max_pos}{The maximum distance from the end of the read upto which mutations are considered.}

\item{flanking_bases}{The number of flanking bases of the mutational signature.}

\item{mutlogo.control}{The control parameters for the mismatch and flanking bases logo.}

\item{breaklogo.control}{The control parameters for the logo for strand break.}

\item{logoport_x}{the X-axis position of the plot window for the logo plot}

\item{logoport_y}{the Y-axis position of the plot window for the logo plot}

\item{logoport_width}{the width of the plot window for the logo plot}

\item{logoport_height}{the width of the plot window for the logo plot}

\item{breaklogoport_x}{the X-axis position of the plot window for strand break logo plot.}

\item{breaklogoport_y}{the Y-axis position of the plot window for the strand break logo plot.}

\item{breaklogoport_width}{the width of the plot window for the strand break logo plot.}

\item{breaklogoport_height}{the width of the plot window for the strand break logo plot.}

\item{lineport_x}{the X-axis position of the plot window for the mutational profile line plot.}

\item{lineport_y}{the Y-axis position of the plot window for the mutational profile line plot.}

\item{lineport_width}{the width of the plot window for the mutational profile line plot.}

\item{lineport_height}{the width of the plot window for the mutational profile line plot.}

\item{output_dir}{The directory where the logo plot will be saved.}

\item{output_width}{The width of the logo plot figure.}

\item{output_height}{the height of the logo plot figure.}
}
\value{
Returns logo plots for each cluster

Returns logo plot for each cluster
}
\description{
Damage Logo plots for each cluster from the GoM model fit. It showcases the
different mutational features - for example, mutation, flanking base, position on read,
strand and strand break information for each cluster.
}
