% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_theta.R
\name{decompose_theta}
\alias{decompose_theta}
\title{Decompose the theta matrix output from model_archaic into probabilities
of different mismatch features}
\usage{
decompose_theta(theta_out, max_pos = 20)
}
\arguments{
\item{theta_out}{A theta matrix from a GoM model fit, with columns
representing clusters and rows representing the mutational
signatures.}

\item{max_pos}{The maximum distance from the end of the read that is used for
filtering.}
}
\value{
Returns a list with number of items equal to number of clusters.
        Each item of this list is another comprising of the probability
        distribution of types of mismatch, flanking base,
        distance from end of read and strand break composition
        separately.
}
\description{
Decomposes the theta matrix outout from \code{model_archaic}
             into component probability distributions for the mismatch type,
             flanking bases, strand break base and position of strand
             composition. These are the probabilities that get represented
             in the logo plot in \code{plot_archaic}.
}
\keyword{decompose_theta}
