% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_mutation.R
\name{filter_by_mutation}
\alias{filter_by_mutation}
\title{Filters only mutations from mutational signatures in a matrix of signature counts.}
\usage{
filter_by_mutation(mat, flanking_bases = 1)
}
\arguments{
\item{mat}{The matrix of counts of signatures in samples.}

\item{flanking_bases}{The number of flanking bases for each mutation. Defaults to 1.}
}
\value{
Returns a new matrix of filtered signatures, where the filtered signatures represent
        just the mutation types (6 in this case C->T, C->A, C->G, T->A, T->C and T->G). The
        other types of mutations are clubbed with above 6. (see \code{club_signature_counts} for more
        details on clubbing)
}
\description{
This function filters the mutations from mutational signatures and generates
             signature counts for the mutations.
}
\keyword{filter_by_mutation}
