% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_mutation_flank.R
\name{filter_by_mutation_flank}
\alias{filter_by_mutation_flank}
\title{Filters mutations and flanking bases from mutational signatures in a
matrix of signature counts.}
\usage{
filter_by_mutation_flank(mat)
}
\arguments{
\item{mat}{The matrix of counts of signatures in samples.}
}
\value{
Returns a new matrix of filtered signatures, where the filtered signatures represent
        mutation and flanking bases.
}
\description{
This function filters the mutations and flanking basesfrom mutational signatures and
             generates signature counts for the filtered signatures in each sample.
}
\keyword{filter_by_mutation_flank}
