% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_mutation_flank_pos.R
\name{filter_by_mutation_flank_pos}
\alias{filter_by_mutation_flank_pos}
\title{Filters mutations, flanking bases and position from ends of read from
mutational signatures in a matrix of signature counts.}
\usage{
filter_by_mutation_flank_pos(mat, max_pos = 20, flanking_bases = 1)
}
\arguments{
\item{mat}{The matrix of counts of signatures in samples.}

\item{max_pos}{The maximum distance from ends of the read that is used for filtering.
Defaults to 20.}

\item{flanking_bases}{The number of flanking bases to the mutation.Defaults to 1}
}
\value{
Returns a new matrix of filtered signatures, where the filtered signatures
        represent mutation, flanking bases and position of mutation from ends of the read.
}
\description{
This function filters the mutations, flanking bases, and position
             of mismatch from ends of the read, from mutational signatures and
             generates signature counts for the filtered signatures in each sample.
}
\keyword{filter_by_mutation_flank_pos}
