% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_pos.R
\name{filter_by_pos}
\alias{filter_by_pos}
\title{Filters position of mutation from ends of read from
mutational signatures in a matrix of signature counts.}
\usage{
filter_by_pos(mat, max_pos = 20)
}
\arguments{
\item{mat}{The matrix of counts of signatures in samples.}

\item{max_pos}{The maximum distance from ends of the read that is used for filtering.
Defaults to 20.}
}
\value{
Returns a new matrix of filtered signatures, where the filtered signatures
        represent position of mutation from ends of the read.
}
\description{
This function filters position of mismatch from ends of the read,
             from mutational signatures and generates signature counts for the filtered
             signatures in each sample.
}
\keyword{filter_by_pos}
