% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_pos_pattern.R
\name{filter_by_pos_pattern}
\alias{filter_by_pos_pattern}
\title{Filters position of a specific type of mutation from ends of read in a matrix of signature counts.}
\usage{
filter_by_pos_pattern(mat, max_pos = 20, pattern = "C->T",
  flanking_bases = 1)
}
\arguments{
\item{mat}{The matrix of counts of signatures in samples.}

\item{max_pos}{The maximum distance from ends of the read that is used for filtering.
Defaults to 20.}

\item{pattern}{The pattern of the mutation type. Defaults to C->T.}

\item{flanking_bases}{the number of flanking bases for the mutation.}
}
\value{
Returns a new matrix of filtered signatures, where the filtered signatures
        represent position of a specific type of mutational type from ends of the read.
}
\description{
This function filters position of a specific type of mismatch
             given by \code{pattern} (for example C->T) from ends of the read,
             and generates signature counts for the filtered signatures in each sample.
}
\keyword{filter_by_pos_pattern}
