% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_out_strand_break.R
\name{filter_out_strand_break}
\alias{filter_out_strand_break}
\title{Filters out strand break information from signatures in a matrix of signature counts.}
\usage{
filter_out_strand_break(mat)
}
\arguments{
\item{mat}{The matrix of counts of signatures in samples.}
}
\value{
Returns a new matrix of filtered signatures, where the filtered signatures
        do not contain the strand break information.
}
\description{
This function filters out strand break information from signatures
             and generates signature counts for the filtered signatures in each sample.
}
\keyword{filter_out_strand_break}
