% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_signatures_only_location.R
\name{filter_signatures_only_location}
\alias{filter_signatures_only_location}
\title{Function to extract signature counts for position along the read}
\usage{
filter_signatures_only_location(mat, max_pos = 20, flanking_bases = 1)
}
\arguments{
\item{mat}{The matrix of the signature counts obtained by \code{aggregate_bin_counts} or
\code{club_signature_counts} functions.}

\item{max_pos}{The highest position along the read from the start for which the counts are to be extracted.}

\item{flanking_bases}{The number of flanking bases in the signatures}
}
\value{
Returns a filtered matrix of signature counts for positions along the read length.
}
\description{
This function can be used by the user to extract the counts data for positions of the mutation signatures
along the read length.
}
\keyword{filter-signatures}
