% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flanking_base_composition.R
\name{flanking_base_composition}
\alias{flanking_base_composition}
\title{Computes flanking base composition for a mutation type in an MFF file.}
\usage{
flanking_base_composition(file, flanking_bases = 1, dist_from_ends = 20,
  pattern = "C->T", plot = TRUE, logo.control = list())
}
\arguments{
\item{file}{an MFF file.}

\item{flanking_bases}{The number of flanking bases in the signatures}

\item{dist_from_ends}{The distance from ends of the reads taken to filter the mutations.
A mutation is not considered if it occurs at a position farther than
this value from either ends of the read.}

\item{pattern}{The type of mutation. The default is C->T.}

\item{plot}{if TRUE, then the function plots the nucleotide composition in a logo plot.
Defaults to TRUE.}

\item{logo.control}{The control parameters for the logo plot representation. Defaults to
\code{list()}}
}
\value{
Returns composition of nucleotides at the left flanking base and the right flanking
        base. If plot is TRUE, then it plots a logo plot representation of the nucleotides.
}
\description{
This function computes the composition of nucleotides at the
left and right flanking bases for a particular mutation type (say "C->T").
}
\keyword{flanking_base_composition}
