% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_archaic.R
\name{model_aRchaic}
\alias{model_aRchaic}
\title{Grade of Membership (GoM) model clustering of aDNA samples using DNA damage patterns}
\usage{
model_aRchaic(dirs, K, tol = 100, labs = NULL, levels = NULL,
  gom_method = "independent", gom.control = list(), output_dir = NULL)
}
\arguments{
\item{K}{the number of clusters to fit to the model.}

\item{tol}{The tolerance level of convergence of the GoM model fit}

\item{labs}{The labels used to group the samples in visualization. May be used to distinguish
samples from different labs, or different library prep.}

\item{levels}{The levels of the \code{labs} vector. Defaults to all unique values of \code{labs} vector.}

\item{gom_method}{The GoM method type. Defaults to \code{independent} model proposed by
Y. Shiraichi and M. Stephens. The other option is to use the \code{full}
model which is uses the \code{maptpx} package by Matt Taddy.}

\item{gom.control}{Control parameters for the GoM model fit.}

\item{output_dir}{The output directory where the model is saved. 
If NULL, it picks the current working directory.}
}
\value{
Fits the GoM model on the aggregated data from \code{process_archaic}
  and then saves the model output as a .RData file. Also outputs model
  assessment scores like the BIC.
}
\description{
Performs GoM model clustering of aDNA samples using DNA damage 
patterns- mutation, flanking base, distance from read end and strand 
break information. The default implementation of this model follows the 
modeling framework of Shiraishi et al (2015).
}
