% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_archaic.R
\name{process_aRchaic}
\alias{process_aRchaic}
\title{Process and aggregate the mismatch feature counts from one or multiple 
MFF files}
\usage{
process_aRchaic(dirs, max_pos = 20, from_scratch = FALSE, delete = FALSE,
  output_rda = TRUE)
}
\arguments{
\item{dirs}{The directory/directories containing the MFF files.}

\item{max_pos}{The maximum position from the ends of the reads for which 
mismatches are considered in aRchaic.}

\item{output_rda}{If non-NULL, the processed data for each
directory in \code{dirs} is saved as a .Rdata file.}
}
\value{
Returns a matrix with rows being the samples (each MFF file), 
columns representing the mismatch signatures (comprising of features like
mismatch type, flanking bases and strand break information). 
The cells contain counts of the number of mutational signatures observed in 
that MFF file.
}
\description{
For each file in the vector of directories \code{dirs}, 
aggregates the mismatch feature counts data from the MFF files
}
\keyword{aggregate_counts}
