% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_length_distribution.R
\name{read_length_distribution}
\alias{read_length_distribution}
\title{Read length distribution plots for an MFF file}
\usage{
read_length_distribution(file, end_break = 5, cols = c("red", "green",
  "blue"), cex_legend = 0.4, title = NULL, cex.main = 1, ylimit = NULL)
}
\arguments{
\item{file}{an MFF file}

\item{end_break}{The position on the read from the end of the fragment such that if
C->T change occurs beyond that end break, it is considered in the third class of
reads (potentially damaged reads)}

\item{cols}{The color profiles of the three plots. Defaults to "red", "green" and "blue".}

\item{cex_legend}{The cex of the legend. Defaults to 0.5.}

\item{ylimit}{The limit of the Y axis.}

\item{cex_main}{The cex of the title of the figure. Defaults to 1.}
}
\value{
A 3-fold density plot of an MFF file under three scenarios - all reads,
reads containing C to T mutations and reads containing C to T towards the ends
of the fragments.
}
\description{
For an MFF file, we plot the read length distribution under three settings -
 all reads, reads containing C to T mutations and reads containing C to T towards the
 ends of the fragments.
}
\keyword{read_length}
