% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_memberships.R
\name{read_memberships}
\alias{read_memberships}
\title{Obtain read level cluster memberships using a fitted GoM model from
a MutSignature file.}
\usage{
read_memberships(fit, reads_file, num_cores = NULL, subset = NULL,
  verbose = TRUE, nostrand = FALSE)
}
\arguments{
\item{fit}{a Grade of Membership model saved under \code{model.rda} produced
as output by \code{aRchaic_cluster()} function. The mutation signature
labels must be provided with the \code{theta} matrix of this object.}

\item{reads_file}{A mutSignature file (.csv) file generated by applying
`generate_summary_bams` on the BAM file.}

\item{num_cores}{The number of cores to be used for parallel processing.
If NULL, all detectable cores are used.}

\item{subset}{The number of reads to be subsetted. Applicable when the user
is interested in learning the cluster memberships of only a handful of reads.}

\item{verbose}{if TRUE, the function returns the total time system time taken.}

\item{nostrand}{if FALSE, the strand information is considered as well
in the computation of the cluster membership probabilities. If TRUE,
the strand information is ignored.}
}
\value{
Produces a matrix with each row containing the cluster memberships
of a read, the mutation signatures observed in that read along with the
read ID.
}
\description{
Performs a read level cluster membership assignment using the
fitted model from \code{aRchiac_cluster()} for each read with mismatch in a
BAM file. The BAM file is first processed to obtain the mutSignature file
which is taken as an input for this function along with a fitted GoM model.
}
\keyword{read_memberships}
