% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_memberships_2.R
\name{read_memberships_2}
\alias{read_memberships_2}
\title{Read level cluster membership assignment from fitted GoM model}
\usage{
read_memberships_2(fit, reads_data, method = "independent", add_sig = TRUE,
  round_off = 5)
}
\arguments{
\item{fit}{a Grade of Membership model saved under \code{model.rda} produced
as output by \code{aRchaic_cluster()} function.}

\item{reads_data}{The new counts matrix of signature patterns for each read
in the BAM file.}

\item{method}{The method used for fitting cluster memberships to the read.
There are four options - \code{independent}, \code{independent-nostrand},
\code{lik} and \code{map}.
The method \code{independent} assumes the topic features (flanking bases,
mutations, strand, strand breaks etc) to be independent and uses the
probability weighting of each feature to determine read memberships.
The \code{independent-nostrand} is same as the \code{independent} method
but does not take account of strand and strand break information.
The method \code{lik} assumes a multinomial likelihood over the mismatch
patterns to determine read memberships.
The method \code{map} uses \code{predict.topics} function from maptpx
package for the prediction. The default choice of method is \code{independent}.}

\item{add_sig}{A Boolean stating whether or not to add the mismatch
signature patterns to the table as an additional column. Defaults to TRUE.}

\item{round_off}{A number stating how many places to round off after decimal
for the cluster membership proportions}
}
\value{
Produces a matrix of the cluster assignments per read, with the reads
marked by read identifiers along the rows and the columns representing the
different clusters. The row sums should sum up to 1.
}
\description{
Performs a read level cluster membership assignment using the
fitted model from \code{aRchiac_cluster()} for each read with mismatch in a
BAM file.
}
\keyword{read_memberships_2}
