% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_breakdown.R
\name{theta_breakdown}
\alias{theta_breakdown}
\title{Breaking down the theta matrix obtained from GoM model by different mutational features.}
\usage{
theta_breakdown(theta_pool, flanking_bases = 1, max_pos = 20)
}
\arguments{
\item{theta_pool}{A theta matrix from a GoM model fit, with columns representing clusters
and rows representing the mutational signatures.}

\item{flanking_bases}{The numbe rof flanking bases to the mutation considered. Defaults to 1.}

\item{max_pos}{The maximum distance from the end of the read that is used for filtering.}
}
\value{
Returns a list with number of items equal to number of clusters. Each item of this
        list is another list which comprises of the individual probability distribution of types of
        mutation, flanking base, strand break, distance from end of read and strand break composition
        separately.
}
\description{
Breaking down a theta matrix into component probabilities for strand composition (
        probability that a mutation comes from positive or negative strand for each cluster
        represented by the theta matrix), strand break composition (probability it is A, C, G or T),
        the distance from the end of the read and the mutation and flanking base.
}
\keyword{theta_breakdown}
