% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_archaic.R
\name{view_archaic}
\alias{view_archaic}
\title{Logo plot representation of an MFF file.}
\usage{
view_archaic(file, logo.control = list(), background = "modern",
  title = NULL, output_file = NULL)
}
\arguments{
\item{file}{a MismatchFeatureFormat (MFF) .csv file.}

\item{logo.control}{The control parameters for the logo plot representation of the sample.
Check the input parameters of \code{damageLogo5()} for the control parameters.}

\item{title}{The name of the sample. If not provided, defaults to the filename.}

\item{output_file}{The filename (with path) of the constructed logo plot.}
}
\value{
Creates a logo plot of the mismatch features for a single
        MFF file.
}
\description{
Creates a logo plot representation of the probability distribution
of the different mismatch features - mismatch, flanking bases, strand break and
position of mismatch - for a single MFF (.csv) file.
}
\keyword{view_archaic}
