% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRchaic_pool.R
\name{aRchaic_pool}
\alias{aRchaic_pool}
\title{Aggregating signature counts from MFF files in one or more folders.}
\usage{
aRchaic_pool(folders, pattern = NULL, breaks = c(-1, seq(1, 20, 1)),
  flanking_bases = 1)
}
\arguments{
\item{folders}{a vector of folder names hosting the MFF files. Each folder may represent
MFF files from same source sequenced and analysed together. (see vignette for
examples)}

\item{pattern}{If the user wants to incorporate specific files in the matrix and not all in
files in the folder, she can put the names as a vector in \code{pattern}.}

\item{breaks}{The breaks used for binning the distance from ends of the reads of the mutations, when
processing the MFF files.}

\item{flanking_bases}{The numbe rof flanking bases to the mutation considered. Defaults to 1.}
}
\value{
The function creates a matrix of counts of MFF samples and the mutational signatures.
        Also if the processed files in a folder are not present, it creates them as .rda
        files in the folders.
}
\description{
Aggregate the signatures into bins for each MFF file in each of the folders
listed in the vector \code{folders} and then create a matrix of counts of signatures, where
each row corresponds to one sample MFF file and each column to one signature (binned from
\code{aggregate_bin_counts ()}).
}
\keyword{aRchaic_pool}
